requirements.txt
attr==0.3.2
charset_normalizer==3.3.2
checks==0.2
ConfigParser==7.0.0
contextlib2==21.6.0
cryptography==42.0.7
Cython==3.0.10
docutils==0.21.2
filelock==3.14.0
HTMLParser==0.0.2
hypothesis==6.102.6
importlib_metadata==7.1.0
ipython==8.12.3
ipywidgets==8.1.2
Jinja2==3.1.3
jnius==1.1.0
keyring==25.2.1
mock==5.1.0
mtrand==0.1
mypy==1.10.0
pickle5==0.0.12
Pillow==10.3.0
protobuf==5.27.0
psutil==5.9.8
pyinstaller==6.5.0
pyOpenSSL==24.1.0
pytest==8.2.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
PyYAML==6.0.1
railroad==0.5.0
redis==5.0.4
setuptools==69.1.1
Sphinx==7.3.7
streamlit==1.34.0
tensorflow==2.16.1
tensorflow_intel==2.16.1
thread==2.0.1
threadpoolctl==3.5.0
tornado==6.4
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1